Move plotting functions into utils file
